Add getGridSize to gameBoard
This function was added to keep object-oriented design consistent.
The grid-size is knowledge that the gameBoard module should have;
if gameLogic needs to know the size of the board, it should ask the
gameBoard.

Remove const ROW_SIZE/COL_SIZE in gameBoard as a result, change
lines that used those constants to getGridSize()