Separate userInput module from gameLogic
This separation was done because a majority of user input's
functionality did not need access to players[]. Therefore coupling
was made explicit for the functions that needed access to players[]
by separating the module, while also creating better encapsulation.

Functions were added to set players[] and get player symbols in
gameLogic in order to help enable the separation.