Finish impossible AI for both first/second turn
Impossible AI uses the minimax algorithm in order to determine what
the best possible moves are for a given board state.

Some degree of randomness is given by giving a 50% consideration to
possible moves when they are both the move. Randomness could be
equally split if these moves were saved and chances were equally
split between each possible best move; but this method was not chosen
due to the additional time that would be needed to save these best
possible moves.